feat: introduce rule for collecting Sass files from external NPM packages (#137)
Introduces a rule that collects direct and transitive Sass files
from external NPM packages, allowing for convenient consumption
of these files in a `sass_binary`/`sass_library` without having
to manually reference all individual files in a NPM package. This
cumbersome and becomes even more complicated if the NPM package
does not provide a Sass bundle, but has multiple Sass files it
imports from. Consumers would not want to determine which files
are transitively loaded and need to be specified (this becomes even
more difficult if the NPM package would then rely on another NPM
package. like it would be case for Angular Material on Angular CDK).